chore(question card): moved ui components to different folder for better structure
