+++ app/client/src/components/ui/Question/TranslatedQuestionContent.tsx
