+++ app/client/src/features/events/Questions/QuestionActions/Quote.tsx
@@ -8,8 +8,7 @@ import type { QuoteMutation } from '@local/__generated__/QuoteMutation.graphql';
 import { useEvent } from '@local/features/events';
 import { ResponsiveDialog, useResponsiveDialog } from '@local/components/ResponsiveDialog';
 import { QuestionForm, QuestionFormProps } from '../QuestionForm';
-import { QuestionAuthor } from '../QuestionAuthor';
-import { QuestionContent } from '../QuestionContent';
+import { QuestionAuthor, QuestionContent } from '@local/components/ui/Question';
 import * as ga from '@local/utils/ga/index';
 import { isURL } from '@local/utils';
 import { QUESTIONS_MAX_LENGTH } from '@local/utils/rules';
