+++ app/client/src/features/events/Questions/index.ts
@@ -1,5 +1,1 @@
-export * from './QuestionAuthor';
-export * from './QuestionContent';
-export * from './QuestionQuote';
-export * from './QuestionStats';
 export * from './QuestionList';
